backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
groq==0.31.1
spacy==3.8.7
psycopg2-binary==2.9.10
redis==6.4.0
python-multipart==0.0.20
python-dotenv==1.1.1
crewai==0.201.1
asyncpg==0.30.0
pytesseract==0.3.13
pymupdf==1.26.4
pdf2image==1.17.0
pillow==12.1.0
python-docx==1.2.0
